requirements/requirements.txt
pandas<2.0
opencv-python==4.6.0.66
scipy==1.9.1
scikit-image==0.19.3
scikit-learn
lightning
pytorch-lightning[extra]
segmentation_models_pytorch
tensorboard
torchvision
loguru
tqdm
prettytable
pydicom
fastparquet
pyarrow
h5py
python-dateutil
tqdm
pylibjpeg
pylibjpeg-openjpeg
pylibjpeg-rle